app/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asn1crypto==1.5.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
bcrypt==4.1.2
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloud-sql-python-connector==1.6.0
colorama==0.4.6
cryptography==42.0.2
dnspython==2.5.0
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.109.0
frozenlist==1.4.1
google-api-core==2.16.2
google-auth==2.27.0
google-cloud-core==2.4.1
google-cloud-secret-manager==2.18.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpc-google-iam-v1==0.13.0
grpcio==1.60.1
grpcio-status==1.60.1
h11==0.14.0
httptools==0.6.1
idna==3.6
jwt==1.3.1
multidict==6.0.5
passlib==1.7.4
pg8000==1.30.4
proto-plus==1.23.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.0
pydantic_core==2.16.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.7
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
scramp==1.4.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
typing_extensions==4.9.0
urllib3==2.2.0
uvicorn==0.27.0.post1
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
